smartnoise/requirements.txt
matplotlib==3.3.1
numpy==1.19.5
opendp-smartnoise==0.1.3.1
pandas==1.2.0
pandasql==0.7.3
scipy==1.5.2
seaborn==0.11.0
psycopg2==2.8.6
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
PyYAML==5.4.1